refactor: :recycle: Polish jest configs
