+++ __tests__/ProcessDetails.spec.tsx
@@ -3,7 +3,7 @@ import { Suspense } from 'react';
 import { render, screen } from '@testing-library/react';
 import { Server } from 'miragejs';
 
-import { setMockUseParams } from '../jest.mock.router';
+import { setMockUseParams } from '../jest.setup';
 import processesData from '../mocks/data/PROCESSES.json';
 import { loadMockServer } from '../mocks/server';
 import { extendedProcessResponse } from '../mocks/server.API';
